Projects/Notes/requirements.txt
# Load .env file
python-dotenv

# fastapi
# sqlalchemy
# uvicorn[standard]


## Django, Postgres, CockroachDB
django
dj_database_url
psycopg2-binary
# django_cockroachdb



# jinja2
# requests
# argparse
# colorlog
# pyopenssl